fixed lost heartbeat test
The stomp.py fixed another issue (https://github.com/jasonrbriggs/stomp.py/pull/234) that hang the main execution awaiting for the heartbeat thread terminate. Now, the heartbeat thread terminate when the main connection closes and the message that we're looking for in the logs is the one indicates it.

We also changed the consumer callback. Now we explicitly await for a heartbeat timeout, instead of only blocking the consumer thread for 3 seconds, that was giving the desired result intermitently.